Move unified plugins to dedicated directory
