+++ scripts/linter/plugins/yaml-loader.js
